+++ android/app/src/main/java/com/boilerplate/app/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.ajeek
+package com.boilerplate.app
 
 import com.facebook.react.ReactActivity
 import com.facebook.react.ReactActivityDelegate
@@ -13,7 +13,7 @@ class MainActivity : ReactActivity() {
    * Returns the name of the main component registered from JavaScript. This is used to schedule
    * rendering of the component.
    */
-  override fun getMainComponentName(): String = "Ajeek"
+  override fun getMainComponentName(): String = "Boiler Plate"
 
   override fun onCreate(savedInstanceState: Bundle?) {
     super.onCreate(null)
