+++ android/app/src/main/java/com/boilerplate/app/MainApplication.kt
@@ -1,4 +1,4 @@
-package com.ajeek
+package com.boilerplate.app
 
 import android.app.Application
 import com.facebook.react.PackageList
