+++ app.json
@@ -1,4 +1,4 @@
 {
-  "name": "Ajeek",
-  "displayName": "Ajeek"
+  "name": "Boiler Plate",
+  "displayName": "Boiler Plate"
 }
