+++ ios/BoilerPlate.xcodeproj/xcshareddata/xcschemes/BoilerPlate.xcscheme
@@ -15,9 +15,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-               BuildableName = "Ajeek.app"
-               BlueprintName = "Ajeek"
-               ReferencedContainer = "container:Ajeek.xcodeproj">
+               BuildableName = "BoilerPlate.app"
+               BlueprintName = "BoilerPlate"
+               ReferencedContainer = "container:BoilerPlate.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -33,9 +33,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "AjeekTests.xctest"
-               BlueprintName = "AjeekTests"
-               ReferencedContainer = "container:Ajeek.xcodeproj">
+               BuildableName = "BoilerPlateTests.xctest"
+               BlueprintName = "BoilerPlateTests"
+               ReferencedContainer = "container:BoilerPlate.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -55,9 +55,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "Ajeek.app"
-            BlueprintName = "Ajeek"
-            ReferencedContainer = "container:Ajeek.xcodeproj">
+            BuildableName = "BoilerPlate.app"
+            BlueprintName = "BoilerPlate"
+            ReferencedContainer = "container:BoilerPlate.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </LaunchAction>
@@ -72,9 +72,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "Ajeek.app"
-            BlueprintName = "Ajeek"
-            ReferencedContainer = "container:Ajeek.xcodeproj">
+            BuildableName = "BoilerPlate.app"
+            BlueprintName = "BoilerPlate"
+            ReferencedContainer = "container:BoilerPlate.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
