+++ ios/BoilerPlateTests/BoilerPlateTests.m
@@ -7,11 +7,11 @@
 #define TIMEOUT_SECONDS 600
 #define TEXT_TO_LOOK_FOR @"Welcome to React"
 
-@interface AjeekTests : XCTestCase
+@interface BoilerPlateTests : XCTestCase
 
 @end
 
-@implementation AjeekTests
+@implementation BoilerPlateTests
 
 - (BOOL)findSubviewInView:(UIView *)view matching:(BOOL (^)(UIView *view))test
 {
