+++ ios/Podfile
@@ -14,7 +14,7 @@ if linkage != nil
   use_frameworks! :linkage => linkage.to_sym
 end
 
-target 'Ajeek' do
+target 'BoilerPlate' do
   config = use_native_modules!
 
   use_react_native!(
@@ -23,7 +23,7 @@ target 'Ajeek' do
     :app_path => "#{Pod::Config.instance.installation_root}/.."
   )
 
-  target 'AjeekTests' do
+  target 'BoilerPlateTests' do
     inherit! :complete
     # Pods for testing
   end
